new setting - force_server_charset This will force the server_charset upon connection open. It is the same as calling mysqli_options on a MS handle, which is the programatical way to do this.
If force_server_charset is used and offline_server_charset
is omitted, internally force_server_charset's value will be used for
offline_server_charset.    
   
If offline_server_charset is set and differs from force_server_charset, there 
will be a fatal error during connect. As this is misconfiguration, fatal is not
too harsh.


git-svn-id: https://svn.php.net/repository/pecl/mysqlnd_ms/trunk@325338 c90b9560-bf6c-de11-be94-00142212c4b1